ensure RelayWrapper contexts get cleaned so store queries aren't left half-read.
supposedly this wasn't a problem because the GC would cleanup these channels that were waiting to be read?
or at least that's what I read, but doesn't look like this was happening.

it would also not be a problem if contexts expire for other reason, but when using background contexts this
was getting ugly.